+++ Blog.DataAccess/Repos/EFPostRepo.cs
@@ -17,7 +17,7 @@ namespace Blog.DataAccess.Entities.Repos
         {
             this.context = context;
         }
-        public void Create(PostInsDTO dto)
+        public void Create(string Content, float Rating, int Views,int UserId)
         {
             context.Posts.Add(new Post()
             {
@@ -30,6 +30,7 @@ namespace Blog.DataAccess.Entities.Repos
 
         public void Delete(int id)
         {
+            if(id!=0)
             context.Posts.Remove(context.Posts.Find(id));
         }
 
@@ -43,22 +44,31 @@ namespace Blog.DataAccess.Entities.Repos
             return context.Posts.Find(id);
         }
 
-        public List<Post> GetPosts(List<int> ids=null)
+        public List<Post> GetPosts(int PageNo, int TagId = 0, int UserId = 0)
         {
-            return ids == null ? context.Posts.ToList() :
-               context.Posts.Where(x => ids.Exists(y => y == x.Id)).ToList();
+            List<Post> list = null;
+            if (TagId == 0 && UserId == 0)
+                list = context.Posts.Where(x => x.Id < PageNo * 5 && x.Id >= (PageNo - 1) * 5).ToList();
+            else if (UserId != 00 && TagId != 0)
+                list = context.Posts.Where(x => x.Id < PageNo * 5 && x.Id >= (PageNo - 1) * 5 && UserId == x.UserId &&
+                  ((List<PostTag>)x.PostTags).Exists(y => y.Id == TagId)).ToList();
+            else if (UserId != 0 && TagId == 0)
+                list = context.Posts.Where(x => x.Id < PageNo * 5 && x.Id >= (PageNo - 1) * 5 && UserId == x.UserId).ToList();
+            else if (UserId == 0 && TagId != 0)
+                list = context.Posts.Where(x => x.Id < PageNo * 5 && x.Id >= (PageNo - 1) * 5 && ((List<PostTag>)x.PostTags).Exists(y => y.Id == TagId)).ToList();
+            return list;
         }
 
-        public void Update(int Id, string Content, int Views,float Rating )
+        public void Update(int Id, string Content, int Views, float Rating)
         {
             Post post = context.Posts.Find(Id);
-            if(String.IsNullOrEmpty(Content))
+            if (String.IsNullOrEmpty(Content))
                 post.Content = Content;
-            if(Views!=0)
-                post.Views = Views; 
+            if (Views != 0)
+                post.Views = Views;
             if (Rating != 0)
                 post.Rating = Rating;
-                
+
             context.Posts.AddOrUpdate(post);
         }
     }
