+++ Blog.DataAccess/Repos/EFUserCredentialRepo.cs
@@ -30,6 +30,7 @@ namespace Blog.DataAccess.Entities.Repos
 
         public void Delete(int id)
         {
+            if(id>0)
             this.context.UserCredentials.Remove(context.UserCredentials.Find(id));
         }
